Update MainActivity.kt
Added a reference to Chronometer and three other buttons. Added running variable to indicate whether the stopwatch is running or not, also an offset variable which is used to display the correct time if the stopwatch is paused or restarted. Then implement the functionality for the stopwatch by setting the onclick listener for the buttons.